chore(processor): move CoCo api client into clients/commercetools folder
