+++ processor/src/server/plugins/ingrid-shipping.plugin.ts
@@ -2,7 +2,7 @@ import { FastifyInstance } from 'fastify';
 import { shippingRoutes } from '../../routes/ingrid-shipping.route';
 
 import { IngridShippingService } from '../../services/ingrid-shipping.service';
-import { CommercetoolsApiClient } from '../../clients/api.client';
+import { CommercetoolsApiClient } from '../../clients/commercetools/api.client';
 import { getConfig } from '../../config/config';
 
 export default async function (server: FastifyInstance) {
